+++ packages/components-vue/src/examples/DrilldownExampleCustomContent.vue
@@ -19,8 +19,8 @@ const onOpen = (): void => {
 const onDismiss = (): void => {
   isDrilldownOpen.value = false;
 };
-const onUpdate = (e: DrilldownUpdateEventDetail): void => {
-  drilldownActiveIdentifier.value = e.activeIdentifier;
+const onUpdate = (e: CustomEvent<DrilldownUpdateEventDetail>): void => {
+  drilldownActiveIdentifier.value = e.detail.activeIdentifier;
 };
 </script>