+++ packages/components/src/components/input-password/input-password.tsx
@@ -141,6 +141,9 @@ export class InputPassword {
 
   @Watch('value')
   public onValueChange(newValue: string): void {
+    if (this.inputElement && this.inputElement.value !== newValue) {
+      this.inputElement.value = newValue;
+    }
     this.internals?.setFormValue(newValue);
   }