+++ packages/components/src/components/input-url/input-url.tsx
@@ -132,6 +132,9 @@ export class InputUrl {
 
   @Watch('value')
   public onValueChange(newValue: string): void {
+    if (this.inputElement && this.inputElement.value !== newValue) {
+      this.inputElement.value = newValue;
+    }
     this.internals?.setFormValue(newValue);
   }